Refactor undo system
